fix(obsd): replace rand() with arc4random() in API tests, set_extent, overhead
On OpenBSD, rand() can return negative values which caused test failures:
- h5_api_test_attribute/dataset/datatype/link/misc/object: negative rand()
  used as dataspace dims/indices in H5_api_test_util.c and
  H5_api_dataset_test.c produced H5S_UNLIMITED or invalid values
- H5TEST-set_extent: negative rand() caused zero chunk dimensions and
  H5S_UNLIMITED dataspace dims in set_extent.c
- PERFORM_overhead: #define rand() ((int)arc4random()) cast uint32_t to
  int, making half of values negative; fixed by removing the (int) cast

Replace rand() with arc4random() in all four files. The arc4random()
compat wrapper in h5test.h provides identical non-negative semantics on
non-OpenBSD platforms via (unsigned int)rand().

+++ tools/test/perform/overhead.c
@@ -32,7 +32,7 @@
 
 /* On OpenBSD, rand() triggers a linker warning suggesting arc4random(). */
 #ifdef __OpenBSD__
-#define rand() ((int)arc4random())
+#define rand() arc4random()
 #endif
 
 #if !defined(H5_HAVE_ATTRIBUTE) || defined __cplusplus
